refactor: move unauthenticated flow from dashboard to web and renamed it to (auth)
