+++ apps/web/app/[locale]/(auth)/components/sign-up.tsx
