+++ apps/web/app/[locale]/(auth)/layout.tsx
