+++ apps/web/app/[locale]/(auth)/sign-in/[[...sign-in]]/page.tsx
@@ -4,37 +4,37 @@ import type { Metadata } from 'next';
 import dynamic from 'next/dynamic';
 import Link from 'next/link';
 
-const title = 'Create an account';
-const description = 'Enter your details to get started.';
-const SignUp = dynamic(() =>
-  import('@/app/(unauthenticated)/components/sign-up').then((mod) => mod.SignUp)
+const title = 'Welcome back';
+const description = 'Enter your details to sign in.';
+const SignIn = dynamic(() =>
+  import('@/app/[locale]/(auth)/components/sign-in').then((mod) => mod.SignIn)
 );
 
 export const metadata: Metadata = createMetadata({ title, description });
 
-const SignUpPage = () => (
+const SignInPage = () => (
   <>
     <div className="flex flex-col space-y-2 text-center">
       <h1 className="font-semibold text-2xl tracking-tight">{title}</h1>
       <p className="text-muted-foreground text-sm">{description}</p>
     </div>
 
-    <SignUp />
+    <SignIn />
 
-    {/* Navigation to sign in */}
+    {/* Navigation to sign up */}
     <div className="text-center">
       <p className="text-muted-foreground text-sm">
-        Already have an account?{' '}
+        Don't have an account?{' '}
         <Button
           variant="link"
           className="h-auto p-0 font-normal text-sm"
           asChild
         >
-          <Link href="/sign-in">Sign in</Link>
+          <Link href="/sign-up">Create an account</Link>
         </Button>
       </p>
     </div>
   </>
 );
 
-export default SignUpPage;
+export default SignInPage;
