+++ apps/web/app/[locale]/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -4,37 +4,37 @@ import type { Metadata } from 'next';
 import dynamic from 'next/dynamic';
 import Link from 'next/link';
 
-const title = 'Welcome back';
-const description = 'Enter your details to sign in.';
-const SignIn = dynamic(() =>
-  import('@/app/(unauthenticated)/components/sign-in').then((mod) => mod.SignIn)
+const title = 'Create an account';
+const description = 'Enter your details to get started.';
+const SignUp = dynamic(() =>
+  import('@/app/[locale]/(auth)/components/sign-up').then((mod) => mod.SignUp)
 );
 
 export const metadata: Metadata = createMetadata({ title, description });
 
-const SignInPage = () => (
+const SignUpPage = () => (
   <>
     <div className="flex flex-col space-y-2 text-center">
       <h1 className="font-semibold text-2xl tracking-tight">{title}</h1>
       <p className="text-muted-foreground text-sm">{description}</p>
     </div>
 
-    <SignIn />
+    <SignUp />
 
-    {/* Navigation to sign up */}
+    {/* Navigation to sign in */}
     <div className="text-center">
       <p className="text-muted-foreground text-sm">
-        Don't have an account?{' '}
+        Already have an account?{' '}
         <Button
           variant="link"
           className="h-auto p-0 font-normal text-sm"
           asChild
         >
-          <Link href="/sign-up">Create an account</Link>
+          <Link href="/sign-in">Sign in</Link>
         </Button>
       </p>
     </div>
   </>
 );
 
-export default SignInPage;
+export default SignUpPage;
